Refactor: change fields in `Atom`

+++ chemrust-core/src/data/custom_data_type/fractional_coord_range.rs
@@ -1,7 +1,7 @@
 use std::fmt::Display;
 
 #[derive(Debug, Clone, Copy)]
-pub struct FractionalCoordRange(f64, f64);
+pub struct FractionalCoordRange(f32, f32);
 
 impl Display for FractionalCoordRange {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
@@ -13,21 +13,21 @@ impl FractionalCoordRange {
     /// Creates a `FractionalCoordRange` with the range restricted to `[0.0, 1.0]`
     /// # Panics
     /// Panics if `lower > higher`
-    pub fn new(lower: f64, higher: f64) -> Self {
+    pub fn new(lower: f32, higher: f32) -> Self {
         assert!(lower <= higher);
         Self(lower.clamp(0.0, 1.0), higher.clamp(0.0, 1.0))
     }
-    pub fn is_in_range(&self, value: f64) -> bool {
+    pub fn is_in_range(&self, value: f32) -> bool {
         if self.0 <= value && self.1 >= value {
             true
         } else {
             false
         }
     }
-    pub fn min(&self) -> f64 {
+    pub fn min(&self) -> f32 {
         self.0
     }
-    pub fn max(&self) -> f64 {
+    pub fn max(&self) -> f32 {
         self.1
     }
 }
